requirements.txt
Streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb